Move productInfoByID from Utils
